hoist common config and informational files to repo root
This PR is a quick quality-of-life improvement that hoists common configuration
out of the individual package directories and into the repository root. This
prevents duplication of common config and readme-style files.

Note that previously the API project was unlicensed, so this PR extends the
scope of the MIT license to cover both codebases.