first pass at typing i18n files
This is a pretty loose first attempt at typing the i18n system. More work needs
to be done to enable strong typing here. Specifically, we need to initialize the
application with a default locale (or interface-compatible placeholder) so that
we can guarantee that there will always be translation functions available in
the I18nContext Provider.