set up modules and ts/yarn config for shared code under packages/common
This change adds support for sharing code between packages and composite
TS compilation.

A lot has changed in the project structure, but both front and backend
can import code from @stateofjs/common (search for that in the project
for a toy implementation of a shared module.)

Still pending some documentation updates, but each individual package
more or less works the same.